#28 added a windows-only distribution script, rearranged files
Added activate.ps1, which creates a distributable orbitx.zip file.
Also moved flight.py up a directory, out of src, and renamed src
to orbitx (so that imports can just be import orbitx.common etc).
Also changed the default behaviour to launch a gui, and renamed
the --gui flag to --no-gui.
And finally, ran autopep8 in preparation for distribution.