#38 Added --profile flag, to generate python-flamegraph reports.
It's too late for me to properly document this, probably a TODO. But
the way to use this flag is run the program with --profile, then use
https://github.com/brendangregg/FlameGraph to read the output log file
and generate a flamegraph.

Also goddang, I really had to do battle with vpython's threading this
time. Fun. But goddang. OrbitX now bypasses vpython's propensity to call
os._exit(0).