#39 added a CLI command to disable sphere segment drawing
If you want to use sphere segments (or just a plain rectangle is fine
for landing view right now), use --sseg to turn them on. If we're using
them long-term, it might be worth it to draw them only when needed, not
all at once on startup.

Also, they crash vpython when the compound object is being created.
I'll ask vpython Bruce why that is.