#27 Factored out a bunch of state-tracking into PhysicsState
PhysicsState wraps both a protos.PhysicalState (naming suggestions
appreciated) and a y_1d. There's a little bit more work to do on it,
e.g. wondering if we should just add .X and .Y and .Spin accessors, or
if there's a more elegant solution.

Also, tests are broken since they use the old Y(y_1d) API.

But for the moment, pushing this commit so Ye Qin can see it and start
working.