#72 Workaround for a vpython bug that made a hyperbolic orbit invisible
Constructing a curve with visible=False and later setting the curve to
visible=True did not properly make the curve visible again, and it seems
that not even setting each point to visible=True made the curve visible.

However, we get the intended behaviour by constructing the curve, then
immediately setting visible=False.