Reconciled landed craft's speed from rotating on earth and relative wind
Now, the habitat landed on Earth and rotating at 450 m/s will be
stationary wrt the wind.

Also, fixed a subtle bug where the physics engine would export a speed
for a landed and rotating habitat that would diverge from the actual
speed that is used inside _derive (which should be the authoritative
value for speed). Added a hopefully explanatory comment about this bug.