#3 This should've been multiple commits i know sorry
First, re-enabled (i.e. removed code rot) for networked communication.
We'll change how the mirror program is launched in the future since I've
had a chance to talk it out with Nova, but the old way of CLI args will
do for now.

Then, I noticed that multiple concurrent copies of orbitx would write to
the same logfile, so I changed the logfile name to include the PID of
the logging process.

Then, I added an atexit handler to delete this logfile when the program
exits cleanly.

Then, I fixed another unrelated exception in the GUI.

TL;DR networking is tested and working again.