Coolant Placement (#131)
* Coolant Window

Moved the coolant and radiators frames to a second tab in the notebook view.

* Added Fuelframe

Added a couple widgets in the grid window

* Rad Shields

Added a new widget for radiation shield in electrical grid. Removed optional_text from some widgets.

* Coolant Window

Moved the coolant and radiators frames to a second tab in the notebook view.

* Rad Shields

Added a new widget for radiation shield in electrical grid. Removed optional_text from some widgets.

* Added widgets to coolant window

* Added a CoolantSection class to organize coolant buttons.